+++ pytest_bdd/types.py
@@ -3,8 +3,10 @@
 FEATURE = 'feature'  # pragma: no cover
 SCENARIO_OUTLINE = 'scenario outline'  # pragma: no cover
 EXAMPLES = 'examples'  # pragma: no cover
+EXAMPLES_VERTICAL = 'examples vertical'  # pragma: no cover
 EXAMPLES_HEADERS = 'example headers'  # pragma: no cover
 EXAMPLE_LINE = 'example line'  # pragma: no cover
+EXAMPLE_LINE_VERTICAL = 'example line vertical'  # pragma: no cover
 SCENARIO = 'scenario'  # pragma: no cover
 GIVEN = 'given'  # pragma: no cover
 WHEN = 'when'  # pragma: no cover