+++ tests/feature/test_steps.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pytest_bdd import scenario, given, when, then
-from pytest_bdd.scenario import GivenAlreadyUsed
+from pytest_bdd import exceptions
 
 
 @scenario('steps.feature', 'Executed step by step')
@@ -84,7 +84,7 @@ def test_multiple_given(request):
     def test():
         pass
 
-    with pytest.raises(GivenAlreadyUsed):
+    with pytest.raises(exceptions.GivenAlreadyUsed):
         test(request)