+++ tests/args/test_args_steps.py
@@ -4,31 +4,46 @@ import functools
 import re
 import pytest
 from pytest_bdd import scenario, given, when, then
-from pytest_bdd.scenario import GivenAlreadyUsed
+from pytest_bdd import exceptions
 
 
-test_steps = scenario(
+@scenario(
     'args_steps.feature',
     'Every step takes a parameter with the same name',
 )
+def test_steps():
+    pass
 
 sc = functools.partial(scenario, 'when_arguments.feature')
-test_argument_in_when_step_1 = sc('Argument in when, step 1')
-test_argument_in_when_step_2 = sc('Argument in when, step 2')
+
+
+@sc('Argument in when, step 1')
+def test_argument_in_when_step_1():
+    pass
+
+
+@sc('Argument in when, step 2')
+def test_argument_in_when_step_2():
+    pass
 
 
 @pytest.fixture
 def values():
-    return ['1', '2', '1', '0', '999999']
+    return [1, 2, 1, 0, 999999]
 
 
-@given(re.compile(r'I have (?P<euro>\d+) Euro'))
+@given(re.compile(r'I have (?P<euro>\d+) Euro'), converters=dict(euro=int))
 def i_have(euro, values):
     assert euro == values.pop(0)
 
 
-@when(re.compile(r'I pay (?P<euro>\d+) Euro'))
+@when(re.compile(r'I pay (?P<euro>\d+) Euro'), converters=dict(euro=int))
 def i_pay(euro, values, request):
+    assert euro == values.pop(0)
+
+
+@then(re.compile(r'I should have (?P<euro>\d+) Euro'), converters=dict(euro=int))
+def i_should_have(euro, values):
     assert euro == values.pop(0)
 
 
@@ -44,11 +59,6 @@ def get_argument(argument, arg):
     argument['arg'] = arg
 
 
-@then(re.compile(r'I should have (?P<euro>\d+) Euro'))
-def i_should_have(euro, values):
-    assert euro == values.pop(0)
-
-
 @then(re.compile('My argument should be (?P<arg>\d+)'))
 def assert_that_my_argument_is_arg(argument, arg):
     """Assert that arg from when equals arg."""
@@ -57,9 +67,11 @@ def assert_that_my_argument_is_arg(argument, arg):
 
 def test_multiple_given(request):
     """Using the same given fixture raises an error."""
-    test = scenario(
+    @scenario(
         'args_steps.feature',
         'Using the same given fixture raises an error',
     )
-    with pytest.raises(GivenAlreadyUsed):
+    def test():
+        pass
+    with pytest.raises(exceptions.GivenAlreadyUsed):
         test(request)