+++ tests/feature/test_scenario.py
@@ -1,15 +1,15 @@
+"""Test scenario decorator."""
 import pytest
 
 from pytest_bdd import scenario
-from pytest_bdd.scenario import ScenarioNotFound
+from pytest_bdd import exceptions
 
 
 def test_scenario_not_found(request):
     """Test the situation when scenario is not found."""
-    test_not_found = scenario(
-        'not_found.feature',
-        'NOT FOUND'
-    )
 
-    with pytest.raises(ScenarioNotFound):
-        test_not_found(request)
+    with pytest.raises(exceptions.ScenarioNotFound):
+        scenario(
+            'not_found.feature',
+            'NOT FOUND'
+        )